Removing the uneeded classes, and adding the setup and frontend class
